Create main_2
this one is only a part of main, keeping this because i find out that tx_desc[n] where n can only be 0 and 1 (in main code, it ranges up until 9. i need to find a way to have that as 9, or maybe have different functions that just use 0 and 1?